# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Wallet Management API
  description: API for managing user wallets, balances, and transactions
  version: 1.0.0
servers:
  - url: https://api.wallet.example.com/v1
    description: Production server
  - url: https://sandbox.wallet.example.com/v1
    description: Sandbox server for testing

tags:
  - name: Wallets
    description: Wallet creation and management
  - name: Balances
    description: Balance retrieval
  - name: Transactions
    description: Deposit, withdrawal, and transfer operations

paths:
  /wallets:
    post:
      tags: [Wallets]
      summary: Create a new wallet
      description: Creates a new wallet for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalletRequest'
      responses:
        '201':
          description: Wallet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          description: Invalid request
        '409':
          description: Wallet already exists for user

  /wallets/{walletId}/balance:
    get:
      tags: [Balances]
      summary: Retrieve current balance
      description: Retrieves the current balance of a user's wallet
      parameters:
        - $ref: '#/components/parameters/walletId'
      responses:
        '200':
          description: Balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '404':
          description: Wallet not found

  /wallets/{walletId}/balance/history:
    get:
      tags: [Balances]
      summary: Retrieve historical balance
      description: Retrieves the balance of a user's wallet at a specific point in the past
      parameters:
        - $ref: '#/components/parameters/walletId'
        - name: timestamp
          in: query
          description: The point in time to check the balance (ISO 8601 format)
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Historical balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '400':
          description: Invalid timestamp format
        '404':
          description: Wallet not found or no data for the specified time

  /wallets/{walletId}/deposit:
    post:
      tags: [Transactions]
      summary: Deposit funds
      description: Deposit money into a wallet
      parameters:
        - $ref: '#/components/parameters/walletId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '200':
          description: Funds deposited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Invalid request
        '404':
          description: Wallet not found

  /wallets/{walletId}/withdraw:
    post:
      tags: [Transactions]
      summary: Withdraw funds
      description: Withdraw money from a wallet
      parameters:
        - $ref: '#/components/parameters/walletId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequest'
      responses:
        '200':
          description: Funds withdrawn successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Invalid request or insufficient funds
        '404':
          description: Wallet not found

  /transfers:
    post:
      tags: [Transactions]
      summary: Transfer funds
      description: Transfer money between user wallets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Funds transferred successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Invalid request or insufficient funds
        '404':
          description: Source or destination wallet not found

components:
  parameters:
    walletId:
      name: walletId
      in: path
      description: Unique identifier of the wallet
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    CreateWalletRequest:
      type: object
      required:
        - userId
        - currency
      properties:
        userId:
          type: string
          description: Unique identifier of the user
          example: "user-12345"
        currency:
          type: string
          description: Currency code (ISO 4217)
          example: "USD"
        initialBalance:
          type: number
          description: Optional initial balance (defaults to 0)
          example: 100.00
          default: 0

    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          example: "user-12345"
        currency:
          type: string
          example: "USD"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"

    Balance:
      type: object
      properties:
        walletId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        amount:
          type: number
          example: 150.75
        currency:
          type: string
          example: "USD"
        asOf:
          type: string
          format: date-time
          description: Timestamp when the balance was recorded
          example: "2023-01-01T12:00:00Z"

    DepositRequest:
      type: object
      required:
        - amount
        - reference
      properties:
        amount:
          type: number
          minimum: 0.01
          example: 50.00
        reference:
          type: string
          description: External reference for the deposit
          example: "deposit-ref-123"
        description:
          type: string
          example: "Bank transfer deposit"

    WithdrawRequest:
      type: object
      required:
        - amount
        - reference
      properties:
        amount:
          type: number
          minimum: 0.01
          example: 25.50
        reference:
          type: string
          description: External reference for the withdrawal
          example: "withdrawal-ref-456"
        description:
          type: string
          example: "ATM withdrawal"

    TransferRequest:
      type: object
      required:
        - sourceWalletId
        - destinationWalletId
        - amount
      properties:
        sourceWalletId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        destinationWalletId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        amount:
          type: number
          minimum: 0.01
          example: 10.00
        reference:
          type: string
          description: Optional reference for the transfer
          example: "transfer-ref-789"
        description:
          type: string
          example: "Payment for services"

    TransactionResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174002"
        walletId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        amount:
          type: number
          example: 50.00
        newBalance:
          type: number
          example: 150.00
        currency:
          type: string
          example: "USD"
        timestamp:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"

    TransferResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174003"
        sourceWalletId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        destinationWalletId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        amount:
          type: number
          example: 10.00
        sourceNewBalance:
          type: number
          example: 90.00
        destinationNewBalance:
          type: number
          example: 110.00
        currency:
          type: string
          example: "USD"
        timestamp:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"